+++ src/Dto/UserBookingStatusInput.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Dto;
+
+class UserBookingStatusInput
+{
+    public array $ids = [];
+}